requirements.txt
python-telegram-bot==21.6
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
alembic==1.13.3
httpx==0.27.2